third_party/WebKit/LayoutTests/fast/animation/request-animation-frame-within-callback-expected.txt

Tests adding one callback within another

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS sameFrame is false
PASS successfullyParsed is true

TEST COMPLETE
